Break out api calls into a separate file
